Remove GH dependence on ScalarTensor
GH shouldn't depend on other evolution systems. The MHD solutions do not depend
on the evolution systems and so this is fine. I think the tags for
ScalarTensor (and probably most of our systems) should be moved to a
PointwiseFunctions/ScalarTensor directory. That's likely where most of the stuff
currently in the system directory should live, with only the assembly for the
executable being in the system. That would include actions,
instantiations, and most of the FD/subcell code since it's not pointwise.